2 dockers et  volume fonctionnels
pb d'appels à la bdd différents selon les fichiers.
      'remplacé
''- DATABASE_URL=sqlite:////db/data.db'''
par - DATABASE_URL=/db/data.db

optimisation du code et nettoyage dans les répertoires.
