+++ docker/docker-compose.yml
@@ -1,20 +1,17 @@
 version: '3.8'
 
 services:
-  
   db:
     image: alpine:latest
     command: /bin/sh -c "apk add --no-cache sqlite && echo 'Executing schema.sql' && sqlite3 /db/data.db '.read /db/schema.sql' && echo 'Schema executed' && tail -f /dev/null"
     volumes:
       - db_data:/db
-      - ./schema.sql:/db/schema.sql
+      - ./db/schema.sql:/db/schema.sql
     healthcheck:
       test: ["CMD-SHELL", "sqlite3 /db/data.db 'SELECT 1'"]
       interval: 10s
       timeout: 5s
       retries: 5
-
-    
 
   app:
     image: python:3.8-slim
@@ -29,7 +26,8 @@ services:
       db:
         condition: service_healthy
     environment:
-      - DATABASE_URL=sqlite:////db/data.db
+
+      - DATABASE_URL=/db/data.db
 
 volumes:
   db_data:
