+++ docker/main.py.asupprimer
@@ -1,0 +1,27 @@
+# main.py
+from fastapi import FastAPI
+from .middleware import log_request_middleware  # Importer le middleware
+import sqlite_utils
+
+db_path = "/db/data.db"
+
+@app.get("/")
+def read_root():
+    return {"Hello": "World"}
+
+app = FastAPI()
+
+# Enregistrer le middleware
+app.middleware("http")(log_request_middleware)
+
+# Vérifiez si la base de données existe, sinon créez-la
+if not os.path.exists(db_path):
+    db = sqlite_utils.Database(db_path)
+    with open("/app/schema.sql", "r") as schema_file:
+        db.executescript(schema_file.read())
+else:
+    db = sqlite_utils.Database(db_path)
+
+# Le reste de votre code FastAPI ici
+
+docker-compose exec app ls -l /db
